+++ src/AI/Node.java
@@ -8,7 +8,7 @@ class Node {
     // node function
     // output = (1/(1 + Math.pow(Math.E, (-1 * output))));
 
-    private double value;
+    private float value;
     private boolean isConstant = false;
 
     Node() {
@@ -20,15 +20,15 @@ class Node {
         value = 1;
     }
 
-    double getValue() {
+    float getValue() {
         return value;
     }
 
     void finalizeValue() {
-        value = (1/(1 + Math.pow(Math.E, (-1 * value))));
+        value = (float) (1/(1 + Math.pow(Math.E, (-1 * value))));
     }
 
-    void updateValue(double newValue) {
+    void updateValue(float newValue) {
         if(!isConstant) {
             value += newValue;
         }
@@ -42,7 +42,7 @@ class Node {
         }
     }
 
-    public void setValue(double newVal) {
+    public void setValue(float newVal) {
         value = newVal;
     }
 }
